refactor(bazaar): split types and builders into http/ and mcp/ directories
Move protocol-specific types and resource service builders into dedicated
subdirectories, mirroring the existing v1/ pattern. The shared types.ts
now contains only union types, the BAZAAR constant, and re-exports.
The shared resourceService.ts is now just the declareDiscoveryExtension
router. No public API changes -- all exports from index.ts are identical.
